renamed files to make them build properly
