+++ src/script/mnpcoinconsensus.h
